requirements.txt
django
djangorestframework
django-environ
mysqlclient
django-mysql
djangorestframework-simplejwt
boto3